Move managers and scripts around
