+++ Assets/_BForBoss/PostProcessingVolume/DeathPostProcessing Profile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e550a5cd92757431491e4bb29aef9f1f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
