+++ Assets/_BForBoss/Prefabs/Checkpoint.prefab
@@ -11,9 +11,6 @@ GameObject:
   - component: {fileID: 1621950425087162391}
   - component: {fileID: 3317328323558938122}
   - component: {fileID: 5588488449876859012}
-  - component: {fileID: 3749082870773717325}
-  - component: {fileID: 6209535401258032501}
-  - component: {fileID: 5468119341624587217}
   m_Layer: 0
   m_Name: Checkpoint
   m_TagString: Untagged
@@ -46,7 +43,7 @@ BoxCollider:
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 6.5, y: 5, z: 1}
+  m_Size: {x: 10, y: 5, z: 4}
   m_Center: {x: 0, y: 0, z: 0}
 --- !u!114 &5588488449876859012
 MonoBehaviour:
@@ -60,79 +57,4 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8798aa975fe401043b93fd9330ce21d0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &3749082870773717325
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3647294160209292200}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 84b4d926caccd644d9c42db18d1f89ff, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  blendMode: 1
-  scaleMode: 0
-  color: {r: 0.44082603, g: 0.6132076, b: 0.31470275, a: 0.34117648}
-  detailLevel: 2
-  renderQueue: -1
-  zTest: 4
-  zOffsetFactor: 0
-  zOffsetUnits: 0
-  stencilComp: 8
-  stencilOpPass: 0
-  stencilRefID: 0
-  stencilReadMask: 255
-  stencilWriteMask: 255
-  size: {x: 6.5, y: 5, z: 1}
-  sizeSpace: 0
---- !u!33 &6209535401258032501
-MeshFilter:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3647294160209292200}
-  m_Mesh: {fileID: 43856717547331454, guid: a3d1f56cebf202c448b6fb479f61a9dc, type: 2}
---- !u!23 &5468119341624587217
-MeshRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3647294160209292200}
-  m_Enabled: 1
-  m_CastShadows: 1
-  m_ReceiveShadows: 1
-  m_DynamicOccludee: 1
-  m_MotionVectors: 1
-  m_LightProbeUsage: 1
-  m_ReflectionProbeUsage: 1
-  m_RayTracingMode: 2
-  m_RayTraceProcedural: 0
-  m_RenderingLayerMask: 1
-  m_RendererPriority: 0
-  m_Materials:
-  - {fileID: 2100000, guid: 0ace0bb3d3bb81c48add355020b81d16, type: 2}
-  m_StaticBatchInfo:
-    firstSubMesh: 0
-    subMeshCount: 0
-  m_StaticBatchRoot: {fileID: 0}
-  m_ProbeAnchor: {fileID: 0}
-  m_LightProbeVolumeOverride: {fileID: 0}
-  m_ScaleInLightmap: 1
-  m_ReceiveGI: 1
-  m_PreserveUVs: 0
-  m_IgnoreNormalsForChartDetection: 0
-  m_ImportantGI: 0
-  m_StitchLightmapSeams: 1
-  m_SelectedEditorRenderState: 3
-  m_MinimumChartSize: 4
-  m_AutoUVMaxDistance: 0.5
-  m_AutoUVMaxAngle: 89
-  m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
-  m_SortingOrder: 0
-  m_AdditionalVertexStreams: {fileID: 0}
+  _showDebugBox: 1
