+++ Assets/_BForBoss/Scripts/Checkpoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8798aa975fe401043b93fd9330ce21d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
