+++ Assets/_BForBoss/Scripts/DeathAreaBehaviour.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace BForBoss
+{
+    public class DeathAreaBehaviour : MonoBehaviour
+    {
+        private void OnCollisionEnter(Collision other)
+        {
+            StateManager.Instance.SetState(State.Death);
+        }
+    }
+}
