+++ Assets/_BForBoss/Scripts/Managers/CheckpointManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26c1a23e1e64fec429a85eb6a194368b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
