+++ Assets/_BForBoss/Scripts/SharedScripts.asmdef
@@ -4,7 +4,9 @@
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f",
         "GUID:31262a2d2355e4cd7a7b39498aa9b78b",
-        "GUID:62883ff587eb248419e6671b1817489b"
+        "GUID:62883ff587eb248419e6671b1817489b",
+        "GUID:b415fcadef204ad4e81c959b085bc506",
+        "GUID:13a6fdc14ca9df84eb8019bc66fa9e5d"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
