+++ Assets/_BForBoss/Scripts/Tags.cs
@@ -1,0 +1,7 @@
+namespace BForBoss
+{
+    public static class Tags
+    {
+        public static string Player => "Player";
+    }
+}
