+++ Assets/_BForBoss/Scripts/Tags.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3fd59734b1368b44a9bbef32480207ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
