--- Assets/_BForBoss/_LevelDesign/Managers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0a97d0db2ce6b9e4bb904baf71289889
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
