--- Assets/_BForBoss/_LevelDesign/Managers/StateManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e4043c29dc413bc43a74490971b9678b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
