--- Assets/_BForBoss/_LevelDesign/Managers/TimeManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b362a49597e2e9b46a9c5be41ef78a96
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
