--- Assets/_BForBoss/_LevelDesign/Managers/WorldManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e1c2a307d30abb641950f9528c996ce8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
