+++ Assets/_BForBoss/_LevelDesign/Prefabs/DeathPlane.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 3662655545432160286}
   - component: {fileID: 2655234651845485709}
   - component: {fileID: 1883633642749820995}
+  - component: {fileID: 429763253822771179}
   m_Layer: 0
   m_Name: DeathPlane
   m_TagString: Untagged
@@ -96,3 +97,15 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: 10209, guid: 0000000000000000e000000000000000, type: 0}
+--- !u!114 &429763253822771179
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5328771532925631528}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2ff954574a9a846dc957686c33e2a3ec, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
