+++ Assets/_BForBoss/_LevelDesign/Scenes/RaceCourse/RaceCourse_GlobalVolumeProfile.asset
@@ -52,43 +52,6 @@ MonoBehaviour:
     m_OverrideState: 0
     m_Value: 15.48
     min: 0
---- !u!114 &-1317665584118534314
-MonoBehaviour:
-  m_ObjectHideFlags: 3
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c5e1dc532bcb41949b58bc4f2abfbb7e, type: 3}
-  m_Name: LensDistortion
-  m_EditorClassIdentifier: 
-  active: 0
-  m_AdvancedMode: 0
-  intensity:
-    m_OverrideState: 1
-    m_Value: 0
-    min: -1
-    max: 1
-  xMultiplier:
-    m_OverrideState: 0
-    m_Value: 1
-    min: 0
-    max: 1
-  yMultiplier:
-    m_OverrideState: 0
-    m_Value: 1
-    min: 0
-    max: 1
-  center:
-    m_OverrideState: 0
-    m_Value: {x: 0.5, y: 0.5}
-  scale:
-    m_OverrideState: 0
-    m_Value: 1
-    min: 0.01
-    max: 5
 --- !u!114 &11400000
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -105,7 +68,6 @@ MonoBehaviour:
   - {fileID: -2558318683599348736}
   - {fileID: 1401687975289329132}
   - {fileID: 2829960594897696778}
-  - {fileID: -1317665584118534314}
 --- !u!114 &1401687975289329132
 MonoBehaviour:
   m_ObjectHideFlags: 3
@@ -122,7 +84,7 @@ MonoBehaviour:
   m_AdvancedMode: 0
   postExposure:
     m_OverrideState: 0
-    m_Value: 0.26
+    m_Value: -0.3
   contrast:
     m_OverrideState: 1
     m_Value: 54.2
@@ -130,7 +92,7 @@ MonoBehaviour:
     max: 100
   colorFilter:
     m_OverrideState: 0
-    m_Value: {r: 0.990566, g: 0.990566, b: 0.990566, a: 1}
+    m_Value: {r: 1, g: 1, b: 1, a: 1}
     hdr: 1
     showAlpha: 0
     showEyeDropper: 1
