--- Assets/_BForBoss/_LevelDesign/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b627937e714b19544944b7e1b1ecf416
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
