--- Assets/_BForBoss/_LevelDesign/Scripts/Tags.cs
@@ -1,7 +1,0 @@
-namespace BForBoss
-{
-    public static class Tags
-    {
-        public static string Player => "Player";
-    }
-}
